+++ packages/common/ProForm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @harbor-design/proform
 
+## 1.4.2
+
+### Patch Changes
+
+- improve required message logic
+
 ## 1.4.1
 
 ### Patch Changes